Renamed file "typeDefs.ts" to "type-defs.ts"
